move some little stuff around
