+++ include/Geode.hpp
@@ -1,5 +1,4 @@
 #include "c++stl/gdstdlib.hpp"
-#include "codegen-base/InterfaceBase.hpp"
 #include "syntax/Modify.hpp"
 #include "syntax/Field.hpp"
 #include "syntax/InternalMacros.hpp"
